✨ feat(): Installed and setup prittier and eslint
