+++ front-end/.prettierrc.json
@@ -1,0 +1,9 @@
+{
+  "parser": "typescript",
+  "printWidth": 100,
+  "singleQuote": false,
+  "trailingComma": "all",
+  "tabWidth": 2,
+  "useTabs": false,
+  "semi": true
+}
