+++ front-end/package.json
@@ -14,15 +14,23 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.3",
-    "react-scripts": "0.9.5"
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
+    "eslint": "^8.27.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.10",
+    "prettier": "2.7.1",
     "typescript": "^4.9.3"
   },
   "scripts": {
-    "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "eslint": "eslint 'src/**/*.{ts, tsx}'",
+    "prettier": "prettier 'src/**/*.{ts, tsx}'",
+    "start": "react-scripts start",
+    "test": "react-scripts test --env=jsdom"
   }
 }
