--- front-end/src/App.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import logo from './logo.svg';
-import './App.css';
-
-export default function App() {
-  return (
-    <div className="App">
-        <div className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h2>Welcome to React</h2>
-        </div>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
-  )
-}
