+++ front-end/src/App.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import "./styles/App.css";
+
+export default function App() {
+  return (
+    <div className="App">
+      <div className="App-header">
+        <h2>Welcome to React</h2>
+      </div>
+      <p className="App-intro">
+        To get started, edit <code>src/App.js</code> and save to reload.
+      </p>
+    </div>
+  );
+}
